+++ src/main/java/site/campingon/campingon/common/jwt/JwtTokenParser.java
@@ -1,0 +1,46 @@
+package site.campingon.campingon.common.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import java.security.Key;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class JwtTokenParser {
+
+    private final Key secretKey;
+
+    public JwtTokenParser(@Value("${jwt.secret-key}") String secretKey) {
+        byte[] keyBytes = secretKey.getBytes();
+        this.secretKey = Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    // Access Token에서 Claims 추출
+    public Claims parseClaims(String accessToken) {
+        try {
+            return Jwts.parserBuilder()
+                .setSigningKey(secretKey)
+                .build()
+                .parseClaimsJws(accessToken)
+                .getBody();
+        } catch (ExpiredJwtException e) {
+            return e.getClaims();
+        }
+    }
+
+    // 토큰에서 이메일 정보 추출
+    public String getEmailFromToken(String accessToken) {
+        Claims claims = Jwts.parserBuilder()
+            .setSigningKey(secretKey)
+            .build()
+            .parseClaimsJws(accessToken)
+            .getBody();
+        return claims.getSubject();
+    }
+
+}
